requirements.txt

requests
python-dotenv
python-csv